+++ packages/expo-updates/src/Updates.ts
@@ -2,6 +2,7 @@ import {
   RCTDeviceEventEmitter,
   CodedError,
   NativeModulesProxy,
+  Platform,
   UnavailabilityError,
 } from '@unimodules/core';
 import { EventEmitter, EventSubscription } from 'fbemitter';
@@ -32,6 +33,15 @@ if (ExpoUpdates.manifestString) {
   _manifest = JSON.parse(ExpoUpdates.manifestString);
 }
 export const manifest: Manifest | object = _manifest ?? {};
+
+if (ExpoUpdates.shouldShowNoRuntimeVersionWarning) {
+  console.warn(
+    'Warning: expo-updates is installed but there is no runtime or SDK version configured. ' +
+      "You'll need to configure one of these two properties in " +
+      Platform.select({ ios: 'Expo.plist', android: 'AndroidManifest.xml' }) +
+      ' before OTA updates will work properly.'
+  );
+}
 
 const isUsingDeveloperTool = !!(manifest as any).developer?.tool;
 const isUsingExpoDevelopmentClient = NativeModulesProxy.ExponentConstants?.appOwnership === 'expo';